cloud-computing/requirements.txt
Flask==3.1.0
pillow==11.0.0
pytest-mock==3.14.0
requests==2.32.3
google-cloud-storage
tensorflow